+++ CHANGELOG.md
@@ -6,9 +6,13 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 - Feature: Add `Pagination` component.
 - Feature: Update styles for `disabled` prop on `ArrowIcon` component.
+- Breaking change: Rename `style` prop to `fontStyle` in `Text` component.
+- Feature: Add `rampProps` prop to `Progress` component.
+- Feature: Add `CrowdfundingCard` component.
 - Feature: Add `mediaQueries` HOC.
 - Fix: Update `primary-light` color.
 - Feature: Add `modifier` prop on `TagButton` component.
+- Breaking change: You must now add a `<StyleRoot>` around `<Hero>` components.
 
 ## [19.5.0] - 2018-02-13